02-VisionApp/CODE/Function-ExtractInfo/requirements.txt
azure-functions>=1.17.0
openai>=1.12.0
pydantic>=2.0.0